CRITICAL: Fix NaN precision issue and comprehensive audit
Phase 7 Started: Fix NaN Precision Issues

CRITICAL BUG FIX:
- Fixed float-to-double conversion in cllm_lattice_cache.c:205
- Changed 'float nan_value = 0.0f / 0.0f' to 'double nan_value = 0.0 / 0.0'
- Ensures consistent double precision throughout system
- Prevents precision loss in NaN markers

COMPREHENSIVE PRECISION AUDIT:
- Scanned entire codebase for float usage
- Found 303 float literals across 20+ files in src/ai/
- Identified major inconsistency:
  * Data structures (AttentionLayer, embeddings, gradients) use double
  * Computation code (attention, training, optimizer) uses float
  * Result: Implicit conversions causing precision loss

CRITICAL FINDINGS:
Priority 1 (Immediate):
- cllm_attention.c: 25 float instances
- cllm_training.c: 27 float instances
- cllm_training_threaded.c: 18 float instances

Priority 2 (High):
- cllm_optimizer.c: 46+17 float instances
- cllm_symmetry.c: 17 float instances
- cllm_root_word_modeling.c: 16 float instances

Total: 303 float literals requiring conversion to double

DOCUMENTATION:
- CRITICAL_NAN_PRECISION_ISSUE.md: Detailed analysis of NaN issue
- PRECISION_AUDIT_RESULTS.md: Complete audit results and fix strategy
- todo.md: Updated with Phase 7 progress and Phase 8 benchmark plan

IMPACT:
- Precision loss: Up to 8 decimal digits in float-to-double conversions
- Training stability: Float precision can cause instability
- Numerical accuracy: Accumulation of rounding errors
- Large models: Float range insufficient

NEXT STEPS:
1. Fix Priority 1 files (attention, training)
2. Test model creation and inference
3. Verify no NaN warnings
4. Continue with Priority 2 files
5. Create comprehensive test suite

Status: Phase 7 in progress - Initial fix complete, systematic fixes needed

+++ src/ai/cllm_lattice_cache.c
@@ -200,7 +200,9 @@ void cllm_embeddings_init_lattice_cached(CLLMModel* model) {
     double* embeddings = model->embeddings.embeddings;
     
     // Mark all embeddings as uninitialized using NaN (0.0/0.0)
-    float nan_value = 0.0f / 0.0f;
+    // CRITICAL FIX: Use double precision for NaN marker (not float)
+    // This ensures consistent precision throughout the system
+    double nan_value = 0.0 / 0.0;
     for (uint32_t i = 0; i < vocab_size * embedding_dim; i++) {
         embeddings[i] = nan_value;  // Special marker for uninitialized
     }
